requirements.txt
streamlit==1.31.1
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
streamlit-authenticator==0.2.3
PyYAML==6.0.1
stripe==7.11.0
python-dotenv==1.0.0
